+++ auditlog/__init__.py
@@ -6,6 +6,3 @@ try:
 except DistributionNotFound:
     # package is not installed
     pass
-
-if django.VERSION < (3, 2):
-    default_app_config = "auditlog.apps.AuditlogConfig"